flytt IBegrunnelse-kode ut i egen fil
